+++ sys-src/backend/app.py
@@ -11,7 +11,6 @@ CORS(app)
 # MongoDB client setup
 client = MongoClient('mongodb', 27017)  # Connect to MongoDB service
 db = client['InfluenzaDB']  # Select database
-registration_collection = db['registration']  # Collection für Registrierungsdaten
 test_collection = db['test']  # Collection for testing
 
 # Define routes
@@ -31,45 +30,6 @@ def test_db():
         return jsonify({"message": "Data inserted successfully!", "id": str(result.inserted_id)})
     except errors.PyMongoError as e:
         app.logger.error(f"Error inserting data into MongoDB: {str(e)}")
-        return make_response(jsonify({"error": str(e)}), 500)
-    
-# Route zum Abrufen aller Registrierungsdaten
-@app.route('/registrations', methods=['GET'])
-def get_registrations():
-    try:
-        data = list(registration_collection.find())
-        for item in data:
-            item["_id"] = str(item["_id"])
-        return jsonify(data)
-    except errors.PyMongoError as e:
-        app.logger.error(f"Error fetching data from MongoDB: {str(e)}")
-        return make_response(jsonify({"error": str(e)}), 500)
-
-# Route zum Aktualisieren von Registrierungsdaten
-@app.route('/register/<id>', methods=['PUT'])
-def update_registration(id):
-    data = request.json
-    try:
-        result = registration_collection.update_one({"_id": ObjectId(id)}, {"$set": data})
-        if result.matched_count > 0:
-            return jsonify({"message": "Update successful"})
-        else:
-            return jsonify({"message": "No document found with that ID"}), 404
-    except errors.PyMongoError as e:
-        app.logger.error(f"Error updating data in MongoDB: {str(e)}")
-        return make_response(jsonify({"error": str(e)}), 500)
-
-# Route zum Löschen von Registrierungsdaten
-@app.route('/register/<id>', methods=['DELETE'])
-def delete_registration(id):
-    try:
-        result = registration_collection.delete_one({"_id": ObjectId(id)})
-        if result.deleted_count > 0:
-            return jsonify({"message": "Delete successful"})
-        else:
-            return jsonify({"message": "No document found with that ID"}), 404
-    except errors.PyMongoError as e:
-        app.logger.error(f"Error deleting data from MongoDB: {str(e)}")
         return make_response(jsonify({"error": str(e)}), 500)
 
 # Run the app
